requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic
langchain
langchain_groq
python-dotenv
opencage